// Code/UBind.FormsApp/src/app/models/field-selectors.enum.ts
export enum FieldSelector {
    Attachment = 'attachment-field',
    Buttons = 'buttons-field',
    Checkbox = 'checkbox-field',
    Content = 'content-field',
    Currency = 'currency-field',
    DatePicker = 'datepicker-field',
    DropDownSelect = 'select-field',
    Hidden = 'hidden-field',
    Iframe = 'iframe-field',
    Password = 'password-field',
    Radio = 'radio-field',
    Repeating = 'repeating-field',
    SearchSelect = 'search-select-field',
    SingleLineText = 'input-field',
    TextArea = 'textarea-field',
    Toggle = 'toggle-field',
    Webhook = 'webhook-field',
    Slider = 'slider-field',
}